split parser and cli
