+++ tests/test_parsers.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 from bidsmreye.bidsmreye import common_parser
+from bidsmreye.download import download_parser
 
 
 def test_parser() -> None:
@@ -46,3 +47,20 @@ def test_parser_basic() -> None:
 
     assert args.task == ["foo", "bar"]
     assert args.non_linear_coreg == False
+
+
+def test_download_parser():
+    parser = download_parser()
+
+    assert parser.description == "Download deepmreye pretrained model from OSF."
+
+    args, _ = parser.parse_known_args(
+        [
+            "--model_name",
+            "1_guided_fixations",
+            "--output_dir",
+            "/home/bob/models",
+        ]
+    )
+
+    assert args.output_dir == "/home/bob/models"
